Add files via upload
Adjus-Table is not just another piece of furniture — it’s a smart space-saving innovation designed for the realities of modern Indian homes. In a country where families share limited living space, Adjus-Table transforms with your needs.
With a simple height-adjustable design, it can instantly switch roles — from a coffee table for your living room, to a study desk for students, a dining surface for family meals, or even a bed table for comfort and convenience.
Our simulation in Autodesk Fusion 360 brings this idea to life, showing how a single, compact structure can deliver functionality, flexibility, and affordability — all in one design.Adjus-Table is more than furniture — it’s a smart solution for small spaces, created with practical thinking and engineering precision.